Time: 3 ms (36.99%) | Memory: 42.2 MB (16.44%) - LeetSync

+++ 238-product-of-array-except-self/product-of-array-except-self.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     vector<int> productExceptSelf(vector<int>& nums) {
         int n = nums.size();
-        vector<long long> pref(n), suf(n);
+        vector<int> pref(n), suf(n);
         pref[0] = suf[n-1] = 1;
         for(int i=1;i<n;i++){
             pref[i] = pref[i-1] * nums[i-1];
